just deleted the unwanted folders
Hello Osuta. Don't mind on the deadline now bro. For just look into the code plus the changes am gonna start applying to it now, but I know by the end of today am gonna do it. Then for the back end, for sure I don't know how to incorporate it there.